vlib:  Fix MacRelix-hosted build errors

+++ v/vlib/vlib/value.cc
@@ -96,7 +96,7 @@ namespace vlib
 		new ((void*) its_box.pointer()) Expr( a, op, b );
 	}
 	
-	Value::Value( size_t n, destructor dtor, value_type t, const dispatch* d )
+	Value::Value( long n, destructor dtor, value_type t, const dispatch* d )
 	:
 		its_box( n, dtor, t )
 	{
