+++ v/vlib/vlib/value.hh
@@ -117,7 +117,7 @@ namespace vlib
 			       const Value&        b,
 			       const dispatch*     d );
 			
-			Value( size_t n, destructor dtor, value_type t, const dispatch* d );
+			Value( long n, destructor dtor, value_type t, const dispatch* d );
 			
 			const void* pointer() const
 			{
